requirements.txt
rembg[cpu]
Pillow